% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_analysis.R
\name{anova_models}
\alias{anova_models}
\title{Run ANOVA tests on all data.}
\usage{
anova_models(data)
}
\arguments{
\item{data}{A dataframe or tibble object containing data in a standardized form. Must include at the minimum: project_name, treatments, replications, variables, value.}
}
\description{
Run ANOVA tests on all data.
}
