% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_cleaning.R
\name{clean_harvester}
\alias{clean_harvester}
\title{Clean harvester (Haldrup) data and put into tidy format. This requires that we know the direction of harvest, the direction of the plot map, and a standardized name for the harvest.}
\usage{
clean_harvester(path)
}
\arguments{
\item{path}{The path to the file in the directory.}
}
\description{
Clean harvester (Haldrup) data and put into tidy format. This requires that we know the direction of harvest, the direction of the plot map, and a standardized name for the harvest.
}
