% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_cleaning.R
\name{clean_nir}
\alias{clean_nir}
\title{Clean NIR data and put into tidy format. This requires that JA's input identifiers as unique and standardized. 
Format: First three letters of the project name, followed by an underscore, followed by the first three letters of the crop, 
followed by an underscore, followed by the treatment number as "T1", followed by the block as "B1". If my project were "Mosaic Potato",
and I were running treatment 1, block 3, the format would be: "Mos_Pot_T1B3".}
\usage{
clean_nir(path, proj_name, label)
}
\arguments{
\item{path}{The path to the file in the directory.}

\item{proj_name}{The name of the project exactly as found in the "Admin Sheet" on OSF.}

\item{label}{The first three letters of the project and the first three letters of the crop. ie "Mos_Pot". Whatever was put into NIR machine.}
}
\description{
Clean NIR data and put into tidy format. This requires that JA's input identifiers as unique and standardized. 
Format: First three letters of the project name, followed by an underscore, followed by the first three letters of the crop, 
followed by an underscore, followed by the treatment number as "T1", followed by the block as "B1". If my project were "Mosaic Potato",
and I were running treatment 1, block 3, the format would be: "Mos_Pot_T1B3".
}
