% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_cleaning.R
\name{clean_potato}
\alias{clean_potato}
\title{Clean potato weights (harvest) data. This requires that the data be in a standardized format which will be determined.}
\usage{
clean_potato(path)
}
\arguments{
\item{path}{The path to the file in the directory.}
}
\description{
Clean potato weights (harvest) data. This requires that the data be in a standardized format which will be determined.
}
