% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_cleaning.R
\name{clean_specgrav}
\alias{clean_specgrav}
\title{Clean specific gravity. This requires that the data be in a standardized format which will be determined.}
\usage{
clean_specgrav(path)
}
\arguments{
\item{path}{The path to the file in the directory.}
}
\description{
Clean specific gravity. This requires that the data be in a standardized format which will be determined.
}
