% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_cleaning.R
\name{compose_data}
\alias{compose_data}
\title{Takes a list of data and puts it together in a single dataframe.}
\usage{
compose_data(data)
}
\arguments{
\item{data}{A list of dataframes.}
}
\description{
Takes a list of data and puts it together in a single dataframe.
}
