% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{download_all_files}
\alias{download_all_files}
\title{Download all files to multiple zip files (currently just .csv data files, but options could be added to allow for download of all types of files) in a Year component.}
\usage{
download_all_files(id = NULL, path)
}
\arguments{
\item{id}{An OSF id to a year component. If no id is specified, will use the current year id.}

\item{path}{Like "path/to/folder" The path to which the file should be downloaded EXCLUDING A FILE NAME. File names will be autogenerated based on project names.}
}
\description{
Download all files to multiple zip files (currently just .csv data files, but options could be added to allow for download of all types of files) in a Year component.
}
