% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{download_files}
\alias{download_files}
\title{Download files from OSF via a specified file ID.}
\usage{
download_files(id, path)
}
\arguments{
\item{id}{An OSF id.}

\item{path}{The path to which the file should be downloaded.}
}
\description{
Download files from OSF via a specified file ID.
}
