% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_admin_sheet.R
\name{get_admin_sheet}
\alias{get_admin_sheet}
\title{Retreive Master Admin Project Sheet}
\usage{
get_admin_sheet(osf_file_id = "z8kpa")
}
\arguments{
\item{osf_file_id}{A string constant containing the code at the end of an OSF url. The default value is the current address to the admin sheet, it might change in the future.
For example, https://osf.io/z8kpa/.
The default string value is directly to the master sheet.}
}
\value{
A tibble in long format
}
\description{
This function retrieves the excel sheet that is updated to contain all current project information. The Admin Sheet is the fulcrum by which the applications operate. If any odd errors arise ensure that you check the admin sheet and that this function is reading it in properly
}
