% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_nodes}
\alias{get_nodes}
\title{Get nodes information from OSF.}
\usage{
get_nodes(id = NULL, contributors = FALSE, files = FALSE,
  children = FALSE)
}
\arguments{
\item{id}{OSF id.}

\item{contributors}{Looking for contributors?}

\item{files}{Looking for files?}

\item{children}{Looking for children...wait...that could be taken the wrong way...Looking for child nodes?}
}
\description{
Get nodes information from OSF.
}
