% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authentication.R
\name{get_osf_pat}
\alias{get_osf_pat}
\title{Retreive Proper OSF PAT from System Environment}
\usage{
get_osf_pat(pat = NULL)
}
\arguments{
\item{pat}{Personal Access Token (PAT) - Create a PAT at https://osf.io/settings/tokens/ if no token is given an error is thrown.}
}
\description{
This function is a replacement for osfr::login() and will properly assign OSF_PAT to the system environment. This is required to make connection with the account on OSF.
}
