% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{move_to_hist}
\alias{move_to_hist}
\title{Function to move files from the Raw_Data component in OSF to the Historical_Raw_Data component so that the file mover for individual projects is faster.}
\usage{
move_to_hist()
}
\description{
Function to move files from the Raw_Data component in OSF to the Historical_Raw_Data component so that the file mover for individual projects is faster.
}
