% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_analysis.R
\name{tukey_models}
\alias{tukey_models}
\title{Run Tukey HSD tests on all data.}
\usage{
tukey_models(data)
}
\arguments{
\item{models}{A list of ANOVA models that is returned by anova_models.}
}
\description{
Run Tukey HSD tests on all data.
}
