% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{upload_files}
\alias{upload_files}
\title{Upload files to OSF via a specified OSF ID.}
\usage{
upload_files(id, path, name = NULL)
}
\arguments{
\item{id}{An OSF id.}

\item{path}{The path to the file to be uploaded.}

\item{name}{The name the file should have on OSF. Defaults to the name of the file in the filepath.}
}
\description{
Upload files to OSF via a specified OSF ID.
}
