% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_analysis.R
\name{write_analysis}
\alias{write_analysis}
\title{Parse Tukey HSD tests to csv and png. (Possibly not png anymore as may not be useful). Will include information on how to read the analysis and what the columns mean, etc.}
\usage{
write_analysis(models)
}
\arguments{
\item{models}{A list of Tukey HSD tibbles that is returned by tukey_models.}
}
\description{
Parse Tukey HSD tests to csv and png. (Possibly not png anymore as may not be useful). Will include information on how to read the analysis and what the columns mean, etc.
}
