% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_analysis.R
\name{anova_models}
\alias{anova_models}
\title{A function to get all ANOVA models. Returns a list of models.}
\usage{
anova_models(data, alpha = 0.05, date = FALSE)
}
\arguments{
\item{data}{A dataframe or tibble object containing data in a standardized form. Must include at the minimum: project_name, treatments, replications, variables, value.}

\item{alpha}{The level of significance}

\item{date}{Whether or not we should split the analysis by date}
}
\description{
A function to get all ANOVA models. Returns a list of models.
}
