% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_analysis.R
\name{write_analysis}
\alias{write_analysis}
\title{Returns an initial report to be reviewed for significance in xlsx format with tabs for each project. Tukey HSD models are written to this report.
Models should be evaluated to determine if orthogonal contrasts should be used, other variables included in the model, or other processes on a case by case basis.}
\usage{
write_analysis(models, path)
}
\arguments{
\item{models}{A list of Tukey HSD tibbles that is returned by tukey_models.}

\item{path}{The path where you want the output saved.}
}
\description{
Returns an initial report to be reviewed for significance in xlsx format with tabs for each project. Tukey HSD models are written to this report.
Models should be evaluated to determine if orthogonal contrasts should be used, other variables included in the model, or other processes on a case by case basis.
}
